refactor: eliminate shell=True and platform.system() branches from subprocess calls
