+++ Repair/repair.py
@@ -7,7 +7,6 @@
 
 import os
 import ast
-import platform
 import subprocess
 from Tool.tool import getAst,getFileName,get_parameter,getLastAPIParameter,resolvePythonExecutable
 from API.LibApi import Parameter
@@ -287,26 +286,19 @@ def validateByRun(callAPI,apiWithValue,projName,virtualEnv,runPath,runCommand):
             l-=1
     pythonPath = resolvePythonExecutable(virtualEnv)
 
-    apiWithValue=apiWithValue.replace('"','\\"')
-    pklPath=pklPath.replace('"','\\"')
-
-    if runPath!='':
-        if platform.system() == "Windows":
-            if runPath not in runCommand:
-                command = f'cd "Dynamic\\{projName}\\{runPath}" && "{pythonPath}" verifySingle.py "{pklPath}" "{apiWithValue}"'
-            else:
-                command=f'cd "Dynamic\\{projName}" && "{pythonPath}" "{runPath}\\verifySingle.py" "{pklPath}" "{apiWithValue}"'
-        else:
-            if runPath not in runCommand:#需要切换到运行文件所在的目录执行命令
-                command=f'cd "Dynamic/{projName}/{runPath}";"{pythonPath}" verifySingle.py "{pklPath}" "{apiWithValue}"'
-            else:
-                command=f'cd "Dynamic/{projName}";"{pythonPath}" "{runPath}/verifySingle.py" "{pklPath}" "{apiWithValue}"'
-    else: #大部分属于这种情况
-        if platform.system() == "Windows":
-            command=f'cd "Dynamic\\{projName}" && "{pythonPath}" verifySingle.py "{pklPath}" "{apiWithValue}"'
-        else:
-            command=f'cd "Dynamic/{projName}";"{pythonPath}" verifySingle.py "{pklPath}" "{apiWithValue}"'
-    result=subprocess.run(command,shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE,text=True)
+    if runPath and runPath not in runCommand:
+        cwd = os.path.join('Dynamic', projName, runPath)
+        script = 'verifySingle.py'
+    elif runPath:
+        cwd = os.path.join('Dynamic', projName)
+        script = os.path.join(runPath, 'verifySingle.py')
+    else:
+        cwd = os.path.join('Dynamic', projName)
+        script = 'verifySingle.py'
+    result = subprocess.run(
+        [pythonPath, script, pklPath, apiWithValue],
+        cwd=cwd, capture_output=True, text=True, encoding='utf-8'
+    )
     return result
